+++ xenon-common/src/main/java/com/vmware/xenon/common/FactoryService.java
@@ -375,6 +375,11 @@ public abstract class FactoryService extends StatelessService {
                 }
             }
 
+            if (!isValidDocumentId(suffix)) {
+                o.fail(new IllegalArgumentException(suffix + ": document id cannot have '/'"));
+                return;
+            }
+
             // check suffix does not already contain the prefix i.e. the factory's self link
             URI serviceUri;
             if (UriUtils.isChildPath(suffix, getSelfLink())) {
@@ -411,6 +416,28 @@ public abstract class FactoryService extends StatelessService {
 
         // complete request, initiate local service start
         completePostRequest(o, childService);
+    }
+
+    private boolean isValidDocumentId(String suffix) {
+        // Skip validation for core services
+        if (suffix.startsWith(ServiceUriPaths.CORE + UriUtils.URI_PATH_CHAR)) {
+            return true;
+        }
+
+        String id = suffix;
+        if (UriUtils.isChildPath(suffix, getSelfLink())) {
+            id = suffix.substring(getSelfLink().length() + 1);
+        }
+
+        if (id.startsWith(UriUtils.URI_PATH_CHAR)) {
+            id = id.substring(1);
+        }
+
+        if (id.contains(UriUtils.URI_PATH_CHAR)) {
+            return false;
+        }
+
+        return true;
     }
 
     protected String buildDefaultChildSelfLink() {